clean structure of row block
